pyproject.toml
[tool.poetry]
name = "streamlit-agent"
version = "0.1.0"
description = ""
authors = ["Tim Conkling <tconkling@gmail.com>", "Joshua Carroll <joshua.carroll@snowflake.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "streamlit_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
streamlit = ">=1.25"
langchain = {extras = ["docarray"], version = "^0.0.234"}
openai = "^0.27.8"
duckduckgo-search = "^3.8.3"
pypdf = "^3.12.2"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
tabulate = "^0.9.0"
pinecone-client = "^2.2.2"
bs4 = "^0.0.1"
nest-asyncio = "^1.5.7"
lxml = "^4.9.3"
tiktoken = "^0.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
